refactor service initializer
