+++ src/decorators/service/service.ts
@@ -1,12 +1,10 @@
-import { container as defaultContainer } from "../container/container";
-import { keys } from "../container/keys";
-import { setGlobal } from "../container/utils/setGlobal";
-import { validateContainer } from "../container/utils/validateContainer";
-import { setInjectables } from "./inject/setInjectables";
-import { Class, ClassDecorator } from "./types";
-import { addClassToContainer } from "./utils/addClassToContainer";
-import { getMetadata } from "./utils/getMetadata";
-import { validateKind } from "./utils/validateKind";
+import { container as defaultContainer } from "../../container/container";
+import { keys } from "../../container/keys";
+import { validateContainer } from "../../container/utils/validateContainer";
+import { getInitializer } from "./getInitializer";
+import { Class, ClassDecorator } from "../types";
+import { addClassToContainer } from "../utils/addClassToContainer";
+import { validateKind } from "../utils/validateKind";
 
 /**
  * A class decorator that makes the class injectable
@@ -44,15 +42,9 @@ export const Service = <T extends Class<object>>(
       throw new Error(`${annotationName} must be used on a named class`);
     }
 
-    context.addInitializer(function () {
-      const service = new this();
-
-      const metadata = getMetadata(annotationName, context);
-      setInjectables(container, service, metadata);
-
-      // @ts-expect-error we need to use object type but it is actually a class
-      setGlobal(container, context.name, service);
-    });
+    context.addInitializer(
+      getInitializer<T>(annotationName, context, container),
+    );
 
     addClassToContainer(container, keys.serviceClasses, constructor);
   }) as ClassDecorator<T>;
